node refactor proposal
